pyproject.toml
[tool.poetry]
name = "sentinel2-background"
version = "0.2.0"
description = ""
authors = ["wwdegroot <wouter.w.de.groot@gmail.com>"]
readme = "README.md"
packages = [{include = "sentinel2_background"}]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
orjson = "^3.9.10"
rich = "^13.7.0"
click = "^8.1.7"
msgspec = "^0.18.4"
asyncclick = "^8.1.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
ruff = "^0.1.5"
pre-commit = "^3.5.0"
pytest-httpx = "^0.27.0"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
